% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Find_nearest_gene.R
\name{Find_nearest_gene}
\alias{Find_nearest_gene}
\title{用于查找SNP附件的基因}
\usage{
Find_nearest_gene(
  data,
  flanking = 0,
  snp = "rsid",
  chr = "chromosome",
  bp = "position",
  build = "hg19",
  filename = "匹配基因文件"
)
}
\arguments{
\item{data}{输入查找的SNP，包含必要的三列SNP,chr.exposure,pos.exposure}

\item{flanking}{输入SNP距离基因的大小}

\item{snp}{输入SNP列名}

\item{chr}{输入染色体列名}

\item{bp}{输入基因组位置}

\item{build}{输入SNP参考的基因组位置，默认hg19，其他hg18或者hg38}

\item{filename}{输出结果的文件夹名称}
}
\description{
用于查找SNP附件的基因
}
