% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finn_read_data.R
\name{finn_read_data}
\alias{finn_read_data}
\title{来源芬兰数据库： https://www.finngen.fi/fi/hyodynna_tuloksia}
\usage{
finn_read_data(file, name, exp_or_out = T, N)
}
\arguments{
\item{file}{下载的文件名称}

\item{name}{输入表型名称}

\item{exp_or_out}{是否作为暴露数据集，默认是T}

\item{N}{样本量大小}
}
\description{
来源芬兰数据库： https://www.finngen.fi/fi/hyodynna_tuloksia
}
