% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giant_read_data.R
\name{giant_read_data}
\alias{giant_read_data}
\title{来源GIANT数据库： https://portals.broadinstitute.org/collaboration/giant/index.php}
\usage{
giant_read_data(file, name, exp_or_out = T)
}
\arguments{
\item{file}{下载的文件名称}

\item{name}{输入表型名称}

\item{exp_or_out}{是否作为暴露数据集，默认是T}
}
\description{
来源GIANT数据库： https://portals.broadinstitute.org/collaboration/giant/index.php
}
