% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWAS_meta.R
\name{GWAS_meta}
\alias{GWAS_meta}
\title{用于两表型的GWAS荟萃分析}
\usage{
GWAS_meta(
  data1,
  data2,
  snp_col = "SNP",
  beta_col = "BETA",
  se_col = "SE",
  a1_col = "effect_allele",
  a2_col = "other_allele",
  filename1 = "芬兰肺癌GWAS",
  filename2 = "catlog肺癌GWAS",
  Nsample = 1e+06,
  triat = "Lung cancer"
)
}
\arguments{
\item{data1}{输入标准MR格式的暴露列名文件，额外包含必要的两列，chr.exposure,pos.exposure}

\item{data2}{输入第二个表型的GWAS文件，需要包含必要的三列，SNP,BETA,SE}

\item{snp_col}{输入SNP列名}

\item{beta_col}{表型2的效应值列名}

\item{se_col}{表型2的效应值标准误列名}

\item{a1_col}{表型2的效应等位基因列名}

\item{a2_col}{表型2的参考等位基因列名}

\item{filename1}{输入临时储存文件夹1}

\item{filename2}{输入临时储存文件夹2}

\item{Nsample}{输入合计的样本量}

\item{triat}{输入表型的名称}
}
\description{
用于两表型的GWAS荟萃分析
}
