% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IEU_0_to_local_2.R
\name{IEU_0_to_local_2}
\alias{IEU_0_to_local_2}
\title{多变量暴露工具变量合并}
\usage{
IEU_0_to_local_2(keyssh, data1IV, data1GWAS, data2IV, data2GWAS)
}
\arguments{
\item{keyssh}{密钥}

\item{data1IV}{来自非IEU暴露1工具变量文件}

\item{data1GWAS}{来自非IEU暴露1的GWAS summary数据文件}

\item{data2IV}{来自非IEU暴露2工具变量文件}

\item{data2GWAS}{来自非IEU暴露2的GWAS summary数据文件}
}
\description{
适用于0个来自IEU的暴露，2个来自非IEU的暴露工具变量合并
}
