% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IEU_0_to_local_5.R
\name{IEU_0_to_local_5}
\alias{IEU_0_to_local_5}
\title{多变量暴露工具变量合并}
\usage{
IEU_0_to_local_5(
  keyssh,
  data1IV,
  data1GWAS,
  data2IV,
  data2GWAS,
  data3IV,
  data3GWAS,
  data4IV,
  data4GWAS,
  data5IV,
  data5GWAS
)
}
\arguments{
\item{keyssh}{密钥}

\item{data1IV}{来自非IEU暴露1工具变量文件}

\item{data1GWAS}{来自非IEU暴露1的GWAS summary数据文件}

\item{data2IV}{来自非IEU暴露2工具变量文件}

\item{data2GWAS}{来自非IEU暴露2的GWAS summary数据文件}

\item{data3IV}{来自非IEU暴露3工具变量文件}

\item{data3GWAS}{来自非IEU暴露3的GWAS summary数据文件}

\item{data4IV}{来自非IEU暴露4工具变量文件}

\item{data4GWAS}{来自非IEU暴露4的GWAS summary数据文件}

\item{data5IV}{来自非IEU暴露4工具变量文件}

\item{data5GWAS}{来自非IEU暴露4的GWAS summary数据文件}
}
\description{
适用于0个来自IEU的暴露，4个来自非IEU的暴露工具变量合并
}
