% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IEU_1_to_local_3.R
\name{IEU_1_to_local_3}
\alias{IEU_1_to_local_3}
\title{多变量暴露工具变量合并}
\usage{
IEU_1_to_local_3(
  keyssh,
  data1IV,
  GWASID,
  data2IV,
  data2GWAS,
  data3IV,
  data3GWAS,
  data4IV,
  data4GWAS
)
}
\arguments{
\item{keyssh}{密钥}

\item{data1IV}{来自IEU暴露1工具变量文件}

\item{GWASID}{来自IEU暴露1的GWAS ID号}

\item{data2IV}{来自非IEU暴露工具变量文件}

\item{data2GWAS}{来自非IEU暴露2的GWAS summary数据文件}

\item{data3IV}{来自非IEU暴露3工具变量文件}

\item{data3GWAS}{来自非IEU暴露3的GWAS summary数据文件}

\item{data4IV}{来自非IEU暴露4工具变量文件}

\item{data4GWAS}{来自非IEU暴露4的GWAS summary数据文件}
}
\description{
适用于1个来自IEU的暴露，3个来自非IEU的暴露工具变量合并
}
