% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IEU_2_to_local_1.R
\name{IEU_2_to_local_1}
\alias{IEU_2_to_local_1}
\title{多变量暴露工具变量合并}
\usage{
IEU_2_to_local_1(
  keyssh,
  data1IV,
  GWASID1,
  data2IV,
  GWASID2,
  data3IV,
  data3GWAS
)
}
\arguments{
\item{keyssh}{密钥}

\item{data1IV}{来自IEU暴露1工具变量文件}

\item{GWASID1}{来自IEU暴露1的GWAS ID号}

\item{data2IV}{来自IEU暴露2工具变量文件}

\item{GWASID2}{来自IEU暴露2的GWAS ID号}

\item{data3IV}{来自非IEU暴露3工具变量文件}

\item{data3GWAS}{来自非IEU暴露3的GWAS summary数据文件}
}
\description{
适用于2个来自IEU的暴露，1个来自非IEU的暴露工具变量合并
}
