% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infla_factor_pre.R
\name{infla_factor_pre}
\alias{infla_factor_pre}
\title{炎症因子数据预处理}
\usage{
infla_factor_pre(
  inputfile,
  sanefile,
  exp_or_out = T,
  P_exp = 1e-05,
  P_out = 5e-08
)
}
\arguments{
\item{inputfile}{炎症因子数据文件位置}

\item{sanefile}{预处理后数据保存位置}

\item{exp_or_out}{处理成暴露还是结局，默认是T表示处理成暴露}

\item{P_exp}{设定处理暴露用于潜在工具变量的P值}

\item{P_out}{设定处理结局用于潜在工具变量的P值}
}
\description{
炎症因子数据预处理
}
