% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metb_pre.R
\name{metb_pre}
\alias{metb_pre}
\title{代谢数据预处理}
\usage{
metb_pre(inputfile, savefile, exp_or_out = T, P_exp = 1e-05, P_out = 5e-08)
}
\arguments{
\item{inputfile}{代谢数据文件位置}

\item{savefile}{预处理后数据保存位置}

\item{exp_or_out}{处理成暴露还是结局，默认是T表示处理成暴露}

\item{P_exp}{设定处理暴露用于潜在工具变量的P值}

\item{P_out}{设定处理结局用于潜在工具变量的P值}
}
\description{
代谢数据预处理
}
