% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Omic_local.R
\name{Omic_local}
\alias{Omic_local}
\title{组学孟德尔随机化}
\usage{
Omic_local(
  keyssh,
  fac_cell_met = 1,
  savefile = "MR结果",
  omicfile,
  finish_out,
  local_clump = F,
  clump_p1 = 1e-05,
  clump_r2 = 0.001,
  clump_kb = 10000,
  pop = "EUR",
  presso = F
)
}
\arguments{
\item{keyssh}{密钥}

\item{fac_cell_met}{选择运行那种组学,1,2,3代表炎症因子,免疫细胞,代谢}

\item{savefile}{结果输出保存的文件夹}

\item{omicfile}{输入组学数据存放的文件夹}

\item{finish_out}{输入结局的GWAS summary}

\item{local_clump}{是否启动本地聚类,默认不启动}

\item{clump_p1}{输入工具变量的选择P值,默认1e-05}

\item{clump_r2}{输入工具变量的选择的r2,默认0.001}

\item{clump_kb}{输入工具变量的选择的距离,默认10000}

\item{pop}{输入工具变量的选择的人群,默认EUR}

\item{presso}{是否启动MRPRESSO,默认F}
}
\description{
组学孟德尔随机化
}
