% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omic_LOO.R
\name{omic_LOO}
\alias{omic_LOO}
\title{组学留一法}
\usage{
omic_LOO(filepath, savefilepath = filepath)
}
\arguments{
\item{filepath}{IV.csv的绝对文件路径例如G:/MR结果}

\item{savefilepath}{保存留一法的绝对路径例如G:/MR结果}
}
\description{
组学留一法
}
