% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLACO_trait.R
\name{PLACO_triat}
\alias{PLACO_triat}
\title{用于两表型的多效性位点分析}
\usage{
PLACO_triat(exp_gwas, out_gwas, p.threshold = 5e-08, save_file = "PLACO")
}
\arguments{
\item{exp_gwas}{暴露的GWAS summary数据}

\item{out_gwas}{结局的GWAS summary数据}

\item{p.threshold}{过滤P值，默认5e-08}

\item{save_file}{输出保存的文件}

\item{keyssh}{学号}
}
\description{
用于两表型的多效性位点分析
}
