% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMR_plot.R
\name{SMR_plot}
\alias{SMR_plot}
\title{适用qtl与表型的SMR分析可视化}
\usage{
SMR_plot(
  keyssh,
  pop = "EUR",
  maGWASfile,
  eqtlfile,
  ILMN_outfile,
  ILMN_ID,
  smr_thresh,
  trait_name,
  num,
  LocusPlot = F,
  EffectPlot = T
)
}
\arguments{
\item{keyssh}{学号}

\item{pop}{参考人种，默认EUR}

\item{maGWASfile}{输入表型gwas数据}

\item{eqtlfile}{输入qtl数据}

\item{ILMN_outfile}{输出特定探针的smr结果}

\item{ILMN_ID}{输入特定探针号}

\item{smr_thresh}{SMR的阈值}

\item{trait_name}{结局名称}

\item{num}{计算机启动线程最大数，默认10}

\item{LocusPlot}{区域GWAS图}

\item{EffectPlot}{SMR散点图}
}
\description{
适用qtl与表型的SMR分析可视化
}
