% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMR_qtl_GWAS.R
\name{SMR_qtl_GWAS}
\alias{SMR_qtl_GWAS}
\title{适用qtl与表型的SMR分析}
\usage{
SMR_qtl_GWAS(
  keyssh,
  pop = "EUR",
  maGWASfile,
  eqtlfile,
  single = T,
  kb = 500,
  r2 = 0.1,
  outfile,
  num = 10
)
}
\arguments{
\item{keyssh}{学号}

\item{pop}{参考人种，默认EUR}

\item{maGWASfile}{输入表型gwas数据}

\item{eqtlfile}{输入qtl数据}

\item{single}{是否top_SNP计算SMR，默认T}

\item{kb}{top_SNP范围}

\item{r2}{top_SNP与区域LD的r2}

\item{outfile}{保存文件名称}

\item{num}{计算机启动线程最大数，默认10}
}
\description{
适用qtl与表型的SMR分析
}
