% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stand_MVMR_IEU_IEU.R
\name{stand_MVMR_IEU_IEU}
\alias{stand_MVMR_IEU_IEU}
\title{基于暴露于结局来自在线数据的标准多变量孟德尔随机化}
\usage{
stand_MVMR_IEU_IEU(
  keyssh,
  exp_GWASID_list = NULL,
  out_GWASID = NULL,
  clump_r2 = 0.001,
  clump_kb = 10000,
  find_proxies = TRUE,
  pval_threshold = 5e-08,
  pop = "EUR",
  ivw_MVMR = T,
  lasso_MVMR = F,
  egger_MVMR = F,
  median_MVMR = F
)
}
\arguments{
\item{keyssh}{密钥}

\item{exp_GWASID_list}{输入一揽子暴露GWAS ID号}

\item{out_GWASID}{输入解决的GWAS ID号}

\item{clump_r2}{输入工具变量的选择的r2,默认0.001}

\item{clump_kb}{输入工具变量的选择的距离,默认10000}

\item{find_proxies}{是否选择代理SNP，默认是TURE}

\item{pval_threshold}{输入工具变量的选择P值,默认5e-08}

\item{pop}{输入人群，默认EUR}

\item{ivw_MVMR}{是否进行IVW方法估计}

\item{lasso_MVMR}{是否进行lasso方法估计}

\item{egger_MVMR}{是否进行egger方法估计}

\item{median_MVMR}{是否进行median方法估计}
}
\description{
基于暴露于结局来自在线数据的标准多变量孟德尔随机化
}
