% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stand_MVMR_local_local.R
\name{stand_MVMR_local_local}
\alias{stand_MVMR_local_local}
\title{基于暴露本地,结局是在线数据的标准多变量孟德尔随机化}
\usage{
stand_MVMR_local_local(
  keyssh,
  exposure_dat_iv,
  outgwas,
  ivw_MVMR = T,
  lasso_MVMR = F,
  egger_MVMR = F,
  median_MVMR = F
)
}
\arguments{
\item{keyssh}{密钥}

\item{exposure_dat_iv}{输入合并的暴露工具变量文件}

\item{outgwas}{数据结局的GWAS summary数据}

\item{ivw_MVMR}{是否进行IVW方法估计}

\item{lasso_MVMR}{是否进行lasso方法估计}

\item{egger_MVMR}{是否进行egger方法估计}

\item{median_MVMR}{是否进行median方法估计}
}
\description{
基于暴露本地,结局是在线数据的标准多变量孟德尔随机化
}
