% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stand_UVMR_IEU_IEU.R
\name{stand_UVMR_IEU_IEU}
\alias{stand_UVMR_IEU_IEU}
\title{适用暴露结局均来自IEU的标准单变量孟德尔随机化分析}
\usage{
stand_UVMR_IEU_IEU(
  keyssh,
  GWASID_exp,
  GWASID_out,
  name_exposure = "exposure",
  samplesize_exposure = 1e+05,
  samplesize_outcome = 1e+05,
  name_outcome = "outcome",
  local_clump = F,
  confounding_SNP = NULL,
  clump_p1 = 5e-08,
  clump_r2 = 0.001,
  clump_kb = 10000,
  pop = "EUR",
  outfile = "MR结果",
  presso = F,
  steiger = T,
  Fvalue = T,
  pt = T
)
}
\arguments{
\item{keyssh}{密钥}

\item{GWASID_exp}{输入暴露的GWAS ID号}

\item{GWASID_out}{输入结局的GWAS ID号}

\item{name_exposure}{输入暴露的名称，默认是exposure}

\item{samplesize_exposure}{输入暴露数据的样本量，默认100000}

\item{samplesize_outcome}{输入结局数据的样本量，默认100000}

\item{name_outcome}{输入结局的名称，默认outcome}

\item{local_clump}{是否启动本地聚类，默认是FALSE}

\item{confounding_SNP}{输入需要去除的混杂因素的rsid}

\item{clump_p1}{输入工具变量的选择P值,默认5e-08}

\item{clump_r2}{输入工具变量的选择的r2,默认0.001}

\item{clump_kb}{输入工具变量的选择的距离,默认10000}

\item{pop}{输入工具变量的选择的人群,默认EUR}

\item{outfile}{输入分析结果的文件夹}

\item{presso}{是否进行MRPRESSO，默认是FALSE}

\item{steiger}{是否进行反向过滤，默认是TURE}

\item{Fvalue}{是否计算F值，默认是TURE}

\item{pt}{是否进行绘图，默认是TURE}
}
\description{
适用暴露结局均来自IEU的标准单变量孟德尔随机化分析
}
