% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stand_UVMR_IEU_local.R
\name{stand_UVMR_IEU_local}
\alias{stand_UVMR_IEU_local}
\title{适用暴露来自IEU结局来自本地数据的标准单变量孟德尔随机化分析}
\usage{
stand_UVMR_IEU_local(
  keyssh,
  GWASID,
  outgwas,
  samplesize_exposure = 1e+05,
  name_exposure = "exposure",
  confounding_search = T,
  local_clump = F,
  confonding_name = NULL,
  clump_p1 = 5e-08,
  clump_r2 = 0.001,
  clump_kb = 10000,
  pop = "EUR",
  outfile = "MR结果",
  presso = F,
  steiger = T,
  Fvalue = T,
  pt = T
)
}
\arguments{
\item{keyssh}{密钥}

\item{GWASID}{输入暴露的GWAS ID号}

\item{outgwas}{输入结局的GWAS摘要数据}

\item{samplesize_exposure}{输入暴露数据的样本量，默认100000}

\item{name_exposure}{输入暴露的名称，默认是exposure}

\item{confounding_search}{是否查询混杂因素，默认是TURE}

\item{local_clump}{是否启动本地聚类，默认是FALSE}

\item{confonding_name}{输入需要去除的混杂因素}

\item{clump_p1}{输入工具变量的选择P值,默认5e-08}

\item{clump_r2}{输入工具变量的选择的r2,默认0.001}

\item{clump_kb}{输入工具变量的选择的距离,默认10000}

\item{pop}{输入工具变量的选择的人群,默认EUR}

\item{outfile}{输入分析结果的文件夹}

\item{presso}{是否进行MRPRESSO，默认是FALSE}

\item{steiger}{是否进行反向过滤，默认是TURE}

\item{Fvalue}{是否计算F值，默认是TURE}

\item{pt}{是否进行绘图，默认是TURE}
}
\description{
适用暴露来自IEU结局来自本地数据的标准单变量孟德尔随机化分析
}
