% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stand_UVMR_local_local.R
\name{stand_UVMR_local_local}
\alias{stand_UVMR_local_local}
\title{适用暴露结局均来自本地数据的标准单变量孟德尔随机化分析}
\usage{
stand_UVMR_local_local(
  keyssh,
  expgwas,
  outgwas,
  local_clump = F,
  confonding_SNP = NULL,
  clump_p1 = 5e-08,
  clump_r2 = 0.001,
  clump_kb = 10000,
  pop = "EUR",
  outfile = "MR结果",
  presso = F,
  steiger = T,
  Fvalue = T,
  pt = T
)
}
\arguments{
\item{keyssh}{密钥}

\item{expgwas}{输入暴露的GWAS摘要数据}

\item{outgwas}{输入结局的GWAS摘要数据}

\item{local_clump}{是否启动本地聚类，默认是FALSE}

\item{confonding_SNP}{输入需要去除的混杂因素}

\item{clump_p1}{输入工具变量的选择P值,默认5e-08}

\item{clump_r2}{输入工具变量的选择的r2,默认0.001}

\item{clump_kb}{输入工具变量的选择的距离,默认10000}

\item{pop}{输入工具变量的选择的人群,默认EUR}

\item{outfile}{输入分析结果的文件夹}

\item{presso}{是否进行MRPRESSO，默认是FALSE}

\item{steiger}{是否进行反向过滤，默认是TURE}

\item{Fvalue}{是否计算F值，默认是TURE}

\item{pt}{是否进行绘图，默认是TURE}
}
\description{
适用暴露结局均来自本地数据的标准单变量孟德尔随机化分析
}
