% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_from_Z.R
\name{trans_from_Z}
\alias{trans_from_Z}
\title{适用根据Z与eaf转换beta和se}
\usage{
trans_from_Z(data, n_col, eaf_col, Z_col, exp_or_out = T)
}
\arguments{
\item{data}{输入需要转换的数据}

\item{n_col}{输入样本量列名}

\item{eaf_col}{输入eaf列名}

\item{Z_col}{输入Z列名}
}
\description{
适用根据Z与eaf转换beta和se
}
