% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_to_eaf.R
\name{trans_to_eaf}
\alias{trans_to_eaf}
\title{适用根据样本量与beta和se转换eaf}
\usage{
trans_to_eaf(data, n_col, beta_col, se_col)
}
\arguments{
\item{data}{输入需要转换的数据}

\item{n_col}{输入样本量列名}

\item{beta_col}{输入效应系数列名}

\item{se_col}{输入标准误列名}
}
\description{
适用根据样本量与beta和se转换eaf
}
