% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yancao_read_data.R
\name{yancao_read_data}
\alias{yancao_read_data}
\title{来源烟草数据库： https://conservancy.umn.edu/handle/11299/201564}
\usage{
yancao_read_data(file, name, exp_or_out = T)
}
\arguments{
\item{file}{下载的文件名称}

\item{name}{输入表型名称}

\item{exp_or_out}{是否作为暴露数据集，默认是T}
}
\description{
来源烟草数据库： https://conservancy.umn.edu/handle/11299/201564
}
